fix: atomic economy flows + NUMERIC/schema drift repair
Economy (collection.controller): wrap combine/open/sell-toys/sell-card/
shop-buy-case/shop-buy-toy/summon-boss in transactions via a new
withTransaction() helper (config/database.js). DELETE ... RETURNING is now
the authoritative concurrency guard (combine/sell), card/toy grants and
catnip spends commit atomically, and catnip is debited via the atomic
spendCatnip() conditional UPDATE everywhere (no more unconditional
addCatnip(-x) that could drive a balance negative). Card/inventory model
fns take an optional executor so they can join a transaction.

Schema drift (migration 023): cat_bosses.week_key DROP NOT NULL — 008
dropped the unique constraint but left NOT NULL, so every boss INSERT
(which passes week_key=NULL) throws on a freshly-migrated DB and no bosses
ever spawn. Also lift gambling_net and boss HP/damage/reward columns from
BIGINT to NUMERIC (matching migration 021's score fix) so fractional or
>9.2e18 deltas no longer throw and abort the score/boss UPDATE. Updated
all ::BIGINT casts on these columns (and the NUMERIC score column in admin
paths) to ::NUMERIC.

All 55 backend tests pass.

+++ src/controllers/boss.controller.js
@@ -189,7 +189,7 @@ const voteBossSpawn = async (req, res, next) => {
 
       const result = await pool.query(
         `INSERT INTO cat_bosses (week_key, boss_name, boss_emoji, max_hp, current_hp, reward_pool, boss_level, buff_duration_minutes, spawn_date, source)
-         VALUES ($1, $2, $3, $4::BIGINT, $4::BIGINT, 0, $5, $6, $7, 'vote')
+         VALUES ($1, $2, $3, $4::NUMERIC, $4::NUMERIC, 0, $5, $6, $7, 'vote')
          ON CONFLICT (spawn_date, boss_name) DO NOTHING RETURNING *`,
         [null, picked.name, picked.emoji, hp, lvl.level, lvl.buffMinutes, dateKey]
       );
